update:#68 メニューにフォーカスを戻すカスタムフック作成
